fix:use initial https git origin and set it to ssh later when keys have been stored

+++ home/.chezmoiscripts/run_once_after_10_postsetup.sh
@@ -2,7 +2,19 @@
 
 if command -v fish >/dev/null; then
   echo install fish plugins
-  for plugin in $(cat "$HOME/.config/fish/fish_plugins"); do
-    fish -c "fisher install '$plugin'"
-  done
+  if test -f "$HOME/.config/fish/fish_plugins"; then
+    for plugin in $(cat "$HOME/.config/fish/fish_plugins"); do
+      fish -c "fisher install '$plugin'"
+    done
+  else
+    echo "$HOME/.config/fish/fish_plugins not found!"
+  fi
 fi
+
+# change chezmoi source to ssh
+curdir=$(pwd)
+path=$(chezmoi source-path)
+cd $path
+git remote set-url origin git@github.com:00sapo/dotfiles.git
+git pull
+cd $curdir
